third_party/blink/web_tests/fast/js/dictionary-string-conversion-exception-expected.txt

Checks that an exception is thrown when a dictionary value cannot be converted to a String

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new MessageEvent('message', {origin: 'test'}).origin is "test"
PASS new MessageEvent('message', {origin: 2}).origin is "2"
PASS new MessageEvent('message', {origin: nonConvertibleToString}) threw exception Exception in toString().
PASS successfullyParsed is true

TEST COMPLETE
